Improve CSystem Init map copy match

+++ src/system.cpp
@@ -598,7 +598,7 @@ void CSystem::Init()
 {
     CFile::CHandle* fileHandle;
     unsigned int count;
-    unsigned int mapSize;
+    int mapSize;
     unsigned int offset;
 
     m_initialized = 1;
@@ -658,7 +658,7 @@ void CSystem::Init()
             for (offset = 0; mapSize != 0; mapSize -= count)
             {
                 count = 0x100000;
-                if (mapSize < 0x100000)
+                if ((unsigned int)mapSize < 0x100000)
                 {
                     count = mapSize;
                 }
